requirements.txt

blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
TheProtocols==3.0.1
urllib3==2.2.2
Werkzeug==3.0.3
BevyFrame @ git+https://github.com/islekcaganmert/bevyframe.git@5bf2347779110ca96b981bbd80fbc483d851a13e